Reorganize patch to de-duplicate code
