+++ troi/patches/similar_recordings.py
@@ -1,98 +1,20 @@
-from datetime import datetime 
-from collections import defaultdict
-from urllib.parse import quote
-import requests
-
 import click
 
-from troi import Element, Artist, Recording, Playlist, PipelineError
-import troi.listenbrainz.stats
 import troi.filters
-import troi.sorts
-import troi.musicbrainz.recording_lookup
-import troi.musicbrainz.recording
+import troi.listenbrainz.stats
 import troi.musicbrainz.mbid_mapping
+import troi.musicbrainz.recording
+import troi.musicbrainz.recording_lookup
+import troi.sorts
+from troi import Recording
+from troi.listenbrainz.similar_recordings import LookupSimilarRecordingsElement
+from troi.filters import FirstArtistCreditFilterElement
 
 
 @click.group()
 def cli():
     pass
 
-class LookupSimilarRecordingsElement(Element):
-    """ Lookup similar recordings.
-
-        Given a list of recordings, this element will output recordings that are similar to the ones passed in.
-    """
-
-    def __init__(self, recording_mbid, alg):
-        super().__init__()
-        self.recording_mbid = recording_mbid
-        self.alg = alg
-
-    @staticmethod
-    def inputs():
-        return []
-
-    @staticmethod
-    def outputs():
-        return [Recording]
-
-    def read(self, inputs):
-
-        sim = [inputs[0][0]]
-        for recording in inputs[0]:
-            url = f"https://labs.api.listenbrainz.org/similar-recordings/json"
-            r = requests.get(url, { "recording_mbid": self.recording_mbid, "algorithm": self.alg })
-            if r.status_code != 200:
-                logger.info("Fetching similar recordings failed: %d. Skipping." % r.status_code)
-                continue
-
-            data = r.json()
-            try:
-                for recording in data[3]["data"]:
-                    sim.append(Recording(mbid=recording["recording_mbid"]))
-            except IndexError:
-                pass
-
-        return sim
-             
-
-class FirstArtistCreditFilterElement(troi.Element):
-    '''
-        Remove all recordings by the artist credit of the FIRST recording in the stream.
-    '''
-
-    @staticmethod
-    def inputs():
-        return [Recording]
-
-    @staticmethod
-    def outputs():
-        return [Recording]
-
-    def read(self, inputs):
-
-        recording = inputs[0][0]
-        acs = recording.artist.mbids
-
-        results = []
-        for i, r in enumerate(inputs[0]):
-            if i == 0:
-                results.append(r)
-                continue
-
-            skip = False
-            for rec_artist in r.artist.mbids:
-                if rec_artist in acs:
-                    skip = True
-                    break
-
-            if skip:
-                continue
-
-            results.append(r)
-
-        return results
 
 class SimilarRecordingsPatch(troi.patch.Patch):
     """
@@ -147,7 +69,7 @@ class SimilarRecordingsPatch(troi.patch.Patch):
 
         recording = troi.musicbrainz.recording.RecordingListElement([Recording(mbid=mbid)])
 
-        similar = LookupSimilarRecordingsElement(mbid, alg)
+        similar = LookupSimilarRecordingsElement(algorithm=alg, count=100)
         similar.set_sources(recording)
 
         recs_lookup = troi.musicbrainz.recording_lookup.RecordingLookupElement()
